Allow admin to add parts with variant selection deferred
This set of changes allows an admin to select "User Selectable" when
adding a part to a product bundle.

Primary use case for this is when a product is part of a bundle and it
only makes sense to allow the User to select a variant rather than
having it be predefined by the admin (e.g. selecting shirt size or
color)

This sets the associated join record (an `AssembliesPart` instance) to
use the master product id for for the selected variant, but also sets
`variant_selection_deferred?` to false so that later we can allow the
front-end to provide a drop-down menu for selecting the desired variant.

- removes `Spree::Product#add_part`
- removes `Spree::Product#set_part_count`
- introduces `Spree::AssignPartToBundleForm` as primary interface for
  assigning parts to bundles (for now)